requirements.txt
numpy==1.22.3
scipy==1.8.0
pandas==1.4.1
matplotlib==3.5.1
scikit-learn==1.0.2
scikit-image==0.19.2
opencv-python==4.8.0.74
Pillow==9.5.0

tqdm==4.63.1
PyYAML==6.0
requests==2.27.1


torch==1.12.0.dev20220327+cu113
torchvision==0.13.0.dev20220327+cu113
torchaudio==0.12.0.dev20220327+cu113

tensorflow-gpu==2.8.0


onnx==1.11.0
onnxruntime-gpu==1.11.0

open3d==0.17.0
pycolmap==0.4.0
pyceres==1.0
kornia==0.6.12

pycocotools==2.0.7
imageio==2.16.1

nuscenes-devkit==1.1.11

omegaconf==2.3.0